pyproject.toml
[project]
name = "mcp-on-cloudrun"
version = "0.1.0"
description = "Example of deploying an MCP server on Cloud Run"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "fastmcp==2.11.1",
    "uvicorn>=0.36.0",
    "google-cloud-storage",
]
